checkShadow splitted
