+++ interface/src/js/checks/checkShadow.js
@@ -1,0 +1,13 @@
+export function checkShadow(name, pokTable) {
+    if (!pokTable[name]) {
+        let index = name.indexOf(" (Shadow)")
+        if (index !== -1) {
+            name = name.slice(0, index)
+            if (!pokTable[name]) {
+                console.log(`Critical: ""${name}" not found in the database`)
+                return ""
+            }
+        }
+    }
+    return name
+}
